rearrange classes into dirs and subdirs
