+++ weaviate/collection/classes/config.py
@@ -1,55 +1,10 @@
-import hashlib
 from dataclasses import dataclass
-from datetime import datetime
 from enum import Enum
-from typing import (
-    Union,
-    Dict,
-    Any,
-    Optional,
-    List,
-    Set,
-    TypeVar,
-    Type,
-    Generic,
-    get_args,
-    get_origin,
-    get_type_hints,
-)
+from typing import Any, Dict, List, Optional, Union
 
-import uuid as uuid_package
-from pydantic import BaseModel, Field, field_validator, ConfigDict
-from pydantic_core._pydantic_core import PydanticUndefined
+from pydantic import BaseModel, Field
 
-from weaviate.util import _to_beacons, _capitalize_first_letter
-from weaviate.weaviate_types import UUID, PYTHON_TYPE_TO_DATATYPE
-from weaviate_grpc import weaviate_pb2
-
-
-@dataclass
-class Error:
-    message: str
-    code: Optional[int] = None
-    original_uuid: Optional[UUID] = None
-
-
-@dataclass
-class _BatchReturn:
-    """This class contains the results of a batch `insert_many` operation.
-
-    Since the individual objects within the batch can error for differing reasons, the data is split up within this class for ease use when performing error checking, handling, and data revalidation.
-
-    Attributes:
-        all_responses: A list of all the responses from the batch operation. Each response is either a `uuid_package.UUID` object or an `Error` object.
-        uuids: A dictionary of all the successful responses from the batch operation. The keys are the indices of the objects in the batch, and the values are the `uuid_package.UUID` objects.
-        errors: A dictionary of all the failed responses from the batch operation. The keys are the indices of the objects in the batch, and the values are the `Error` objects.
-        has_errors: A boolean indicating whether or not any of the objects in the batch failed to be inserted. If this is `True`, then the `errors` dictionary will contain at least one entry.
-    """
-
-    all_responses: List[Union[uuid_package.UUID, Error]]
-    uuids: Dict[int, uuid_package.UUID]
-    errors: Dict[int, Error]
-    has_errors: bool = False
+from weaviate.util import _capitalize_first_letter
 
 
 class DataType(str, Enum):
@@ -352,22 +307,6 @@ class _ReferenceDataTypeMultiTarget:
     target_collections: List[str]
 
 
-class ReferenceDataType(BaseModel):
-    """Use this class when defining the collections to which your collection should cross reference when specifying
-    a reference property.
-
-    A cross reference can refer to multiple collections at once, hence the ability to specify a list of collections.
-
-    However, be careful when doing so as all downstream inserts, queries, and searches for this collection must specify
-    the exact collection to which the reference refers.
-
-    If you are unsure, it is recommended to use a single collection
-    for each cross reference.
-    """
-
-    collections: Union[str, List[str]]
-
-
 @dataclass
 class _Property:
     data_type: Union[DataType, _ReferenceDataType, _ReferenceDataTypeMultiTarget]
@@ -592,6 +531,22 @@ class Property(ConfigCreateModel):
         return ret_dict
 
 
+class ReferenceDataType(BaseModel):
+    """Use this class when defining the collections to which your collection should cross reference when specifying
+    a reference property.
+
+    A cross reference can refer to multiple collections at once, hence the ability to specify a list of collections.
+
+    However, be careful when doing so as all downstream inserts, queries, and searches for this collection must specify
+    the exact collection to which the reference refers.
+
+    If you are unsure, it is recommended to use a single collection
+    for each cross reference.
+    """
+
+    collections: Union[str, List[str]]
+
+
 class ReferencePropertyBase(ConfigCreateModel):
     name: str
 
@@ -662,474 +617,3 @@ class CollectionConfig(CollectionConfigCreateBase):
             ret_dict["properties"] = [prop.to_dict() for prop in self.properties]
 
         return ret_dict
-
-
-class IncludesModel(BaseModel):
-    def to_include(self) -> str:
-        include: List[str] = []
-        for field, value in self:
-            if value:
-                include.append(field)
-        return ",".join(include)
-
-
-class GetObjectByIdIncludes(IncludesModel):
-    classification: bool = False
-    vector: bool = False
-
-
-class GetObjectsIncludes(IncludesModel):
-    classification: bool = False
-    featureProjection: bool = Field(False, alias="feature_projection")
-    vector: bool = False
-
-
-@dataclass
-class _MetadataReturn:
-    # uuid: Optional[uuid_package.UUID] = Field(None, alias="id")
-    # vector: Optional[List[float]] = None
-    # creation_time_unix: Optional[int] = Field(None, alias="creationTimeUnix")
-    # last_update_time_unix: Optional[int] = Field(None, alias="lastUpdateTimeUnix")
-    # distance: Optional[float] = None
-    # certainty: Optional[float] = None
-    # score: Optional[float] = None
-    # explain_score: Optional[str] = Field(None, alias="explainScore")
-    # is_consistent: Optional[bool] = Field(None, alias="isConsistent")
-    uuid: Optional[uuid_package.UUID] = None
-    vector: Optional[List[float]] = None
-    creation_time_unix: Optional[int] = None
-    last_update_time_unix: Optional[int] = None
-    distance: Optional[float] = None
-    certainty: Optional[float] = None
-    score: Optional[float] = None
-    explain_score: Optional[str] = None
-    is_consistent: Optional[bool] = None
-
-
-Properties = TypeVar("Properties")
-
-
-@dataclass
-class _Object(Generic[Properties]):
-    data: Properties
-    metadata: _MetadataReturn
-
-
-def _metadata_from_dict(metadata: Dict[str, Any]) -> _MetadataReturn:
-    return _MetadataReturn(
-        uuid=uuid_package.UUID(metadata["id"]) if "id" in metadata else None,
-        vector=metadata.get("vector"),
-        creation_time_unix=metadata.get("creationTimeUnix"),
-        last_update_time_unix=metadata.get("lastUpdateTimeUnix"),
-        distance=metadata.get("distance"),
-        certainty=metadata.get("certainty"),
-        explain_score=metadata.get("explainScore"),
-        score=metadata.get("score"),
-        is_consistent=metadata.get("isConsistent"),
-    )
-
-
-@dataclass
-class ReferenceTo:
-    uuids: Union[List[UUID], UUID]
-
-    @property
-    def uuids_str(self) -> List[str]:
-        if isinstance(self.uuids, list):
-            return [str(uid) for uid in self.uuids]
-        else:
-            return [str(self.uuids)]
-
-    def to_beacons(self) -> List[Dict[str, str]]:
-        return _to_beacons(self.uuids)
-
-
-@dataclass
-class ReferenceToMultiTarget(ReferenceTo):
-    target_collection: str
-
-    def to_beacons(self) -> List[Dict[str, str]]:
-        return _to_beacons(self.uuids, self.target_collection)
-
-
-@dataclass
-class CrossReference:
-    ref_type: Union[Type, str]
-
-    @property
-    def name(self) -> str:
-        if isinstance(self.ref_type, type):
-            return _capitalize_first_letter(self.ref_type.__name__)
-        else:
-            assert isinstance(self.ref_type, str)
-            return _capitalize_first_letter(self.ref_type)
-
-
-@dataclass
-class BatchReference:
-    from_uuid: UUID
-    to_uuid: UUID
-
-
-@dataclass
-class DataObject:
-    data: Dict[str, Any]
-    uuid: Optional[UUID] = None
-    vector: Optional[List[float]] = None
-
-
-class BaseProperty(BaseModel):
-    uuid: uuid_package.UUID = Field(default_factory=uuid_package.uuid4)
-    vector: Optional[List[float]] = None
-
-    # def __new__(cls, *args, **kwargs):
-    #     #
-    #     build = super().__new__(cls)
-    #     # fields, class_vars = collect_model_fields(cls)
-    #     for name, field in build.model_fields.items():
-    #         if name not in BaseProperty.model_fields:
-    #             field_type = build._remove_optional_type(field.annotation)
-    #             if inspect.isclass(field_type):
-    #                 if field.annotation not in PYTHON_TYPE_TO_DATATYPE:
-    #                     build.model_fields[name] = fields.FieldInfo(annotation=typing.Optional[UUID], default=None)
-    #
-    #     build.__class_vars__.update(build.__class_vars__)
-    #     return build
-    #
-    #
-    # make references optional by default - does not work
-
-    def model_post_init(self, __context: Any) -> None:
-        self._reference_fields: Set[str] = self.get_ref_fields(type(self))
-
-        self._reference_to_class: Dict[str, str] = {}
-        for ref in self._reference_fields:
-            self._reference_to_class[ref] = self.model_fields[ref].metadata[0].name
-
-    @staticmethod
-    def get_ref_fields(model: Type["BaseProperty"]) -> Set[str]:
-        return {
-            name
-            for name, field in model.model_fields.items()
-            if (
-                field.metadata is not None
-                and len(field.metadata) > 0
-                and isinstance(field.metadata[0], CrossReference)
-            )
-            and name not in BaseProperty.model_fields
-        }
-
-    @staticmethod
-    def get_non_ref_fields(model: Type["BaseProperty"]) -> Set[str]:
-        return {
-            name
-            for name, field in model.model_fields.items()
-            if (
-                field.metadata is None
-                or len(field.metadata) == 0
-                or isinstance(field.metadata[0], PropertyConfig)
-            )
-            and name not in BaseProperty.model_fields
-        }
-
-    def props_to_dict(self, update: bool = False) -> Dict[str, Any]:
-        fields_to_exclude: Set[str] = self._reference_fields.union({"uuid", "vector"})
-        if update:
-            fields_to_exclude.union(
-                {field for field in self.model_fields.keys() if field not in self.model_fields_set}
-            )
-
-        c = self.model_dump(exclude=fields_to_exclude)
-        for ref in self._reference_fields:
-            val = getattr(self, ref, None)
-            if val is not None:
-                c[ref] = _to_beacons(val, self._reference_to_class[ref])
-        return c
-
-    @field_validator("uuid")
-    def create_valid_uuid(cls, input_uuid: UUID) -> uuid_package.UUID:
-        if isinstance(input_uuid, uuid_package.UUID):
-            return input_uuid
-
-        # see if str is already a valid uuid
-        try:
-            return uuid_package.UUID(input_uuid)
-        except ValueError:
-            hex_string = hashlib.md5(input_uuid.encode("UTF-8")).hexdigest()
-            return uuid_package.UUID(hex=hex_string)
-
-    @staticmethod
-    def type_to_dict(model: Type["BaseProperty"]) -> List[Dict[str, Any]]:
-        types = get_type_hints(model)
-
-        non_optional_types = {
-            name: BaseProperty.remove_optional_type(tt)
-            for name, tt in types.items()
-            if name not in BaseProperty.model_fields
-        }
-
-        non_ref_fields = model.get_non_ref_fields(model)
-        properties = []
-        for name in non_ref_fields:
-            prop = {
-                "name": name,
-                "dataType": [PYTHON_TYPE_TO_DATATYPE[non_optional_types[name]]],
-            }
-            metadata_list = model.model_fields[name].metadata
-            if metadata_list is not None and len(metadata_list) > 0:
-                metadata = metadata_list[0]
-                if isinstance(metadata, PropertyConfig):
-                    prop.update(metadata.to_dict())
-
-            properties.append(prop)
-
-        reference_fields = model.get_ref_fields(model)
-        properties.extend(
-            {
-                "name": name,
-                "dataType": [model.model_fields[name].metadata[0].name],
-            }
-            for name in reference_fields
-        )
-
-        return properties
-
-    @staticmethod
-    def type_to_properties(
-        model: Type["BaseProperty"],
-    ) -> List[Union[Property, ReferenceProperty, ReferencePropertyMultiTarget]]:
-        types = get_type_hints(model)
-
-        non_optional_types = {
-            name: BaseProperty.remove_optional_type(tt)
-            for name, tt in types.items()
-            if name not in BaseProperty.model_fields
-        }
-
-        non_ref_fields = model.get_non_ref_fields(model)
-        properties: List[Union[Property, ReferenceProperty, ReferencePropertyMultiTarget]] = []
-        for name in non_ref_fields:
-            data_type = [PYTHON_TYPE_TO_DATATYPE[non_optional_types[name]]]
-            prop: Dict[str, Any] = {}
-            metadata_list = model.model_fields[name].metadata
-            if metadata_list is not None and len(metadata_list) > 0:
-                metadata = metadata_list[0]
-                if isinstance(metadata, PropertyConfig):
-                    prop.update(metadata.to_dict())
-
-            properties.append(Property(name=name, data_type=DataType(data_type[0]), **prop))
-
-        reference_fields = model.get_ref_fields(model)
-        properties.extend(
-            ReferenceProperty(
-                name=name,
-                target_collection=model.model_fields[name].metadata[0].name,
-            )
-            for name in reference_fields
-        )
-
-        return properties
-
-    @staticmethod
-    def get_non_default_fields(model: Type["BaseProperty"]) -> Set[str]:
-        return {
-            field
-            for field, val in model.model_fields.items()
-            if val.default == PydanticUndefined and field not in BaseProperty.model_fields.keys()
-        }
-
-    @staticmethod
-    def remove_optional_type(python_type: type) -> type:
-        is_list = get_origin(python_type) == list
-        args = get_args(python_type)
-        if len(args) == 0:
-            return python_type
-
-        return_type = [t for t in args if t is not None][0]
-
-        if is_list:
-            return List[return_type]
-        else:
-            return return_type
-
-
-Model = TypeVar("Model", bound=BaseProperty)
-
-
-class RefToObjectModel(BaseModel, Generic[Model]):
-    uuids_to: Union[List[UUID], UUID] = Field()
-
-    def to_beacon(self) -> List[Dict[str, str]]:
-        return _to_beacons(self.uuids_to)
-
-
-UserModelType = Type[BaseProperty]
-
-
-class CollectionModelConfig(CollectionConfigCreateBase, Generic[Model]):
-    model: Type[Model]
-
-    def to_dict(self) -> Dict[str, Any]:
-        ret_dict = super().to_dict()
-
-        ret_dict["class"] = _capitalize_first_letter(self.model.__name__)
-
-        if self.model is not None:
-            ret_dict["properties"] = self.model.type_to_dict(self.model)
-
-        return ret_dict
-
-
-class TenantActivityStatus(str, Enum):
-    """TenantActivityStatus class used to describe the activity status of a tenant in Weaviate.
-
-    Attributes
-    HOT: The tenant is fully active and can be used.
-    COLD: The tenant is not active, files stored locally.
-    """
-
-    HOT = "HOT"
-    COLD = "COLD"
-
-
-class Tenant(BaseModel):
-    """Tenant class used to describe a tenant in Weaviate.
-
-    Attributes
-    name: the name of the tenant.
-    activity_status : TenantActivityStatus, default: "HOT"
-    """
-
-    model_config = ConfigDict(populate_by_name=True)
-    name: str
-    activityStatus: TenantActivityStatus = Field(
-        default=TenantActivityStatus.HOT, alias="activity_status"
-    )
-
-    @property
-    def activity_status(self) -> TenantActivityStatus:
-        return self.activityStatus
-
-
-class _Filters:
-    def __and__(self, other: "_Filters"):
-        return _FilterAnd(self, other)
-
-    def __or__(self, other: "_Filters"):
-        return _FilterOr(self, other)
-
-
-class _FilterAnd(_Filters):
-    def __init__(self, *args: _Filters):
-        self.filters: List[_Filters] = list(args)
-
-    # replace with the following once 3.11 is the minimum version
-    #     Operator: weaviate_pb2.Filters.OperatorType = weaviate_pb2.Filters.OperatorAnd
-    @property
-    def operator(self) -> weaviate_pb2.Filters.Operator:
-        return weaviate_pb2.Filters.OPERATOR_AND
-
-
-class _FilterOr(_Filters):
-    def __init__(self, *args: _Filters):
-        self.filters: List[_Filters] = list(args)
-
-    # replace with the following once 3.11 is the minimum version
-    #     Operator: weaviate_pb2.Filters.OperatorType = weaviate_pb2.Filters.OperatorOr
-    @property
-    def operator(self) -> weaviate_pb2.Filters.Operator:
-        return weaviate_pb2.Filters.OPERATOR_OR
-
-
-FilterValuesList = Union[List[str], List[bool], List[int], List[float], List[datetime.date]]
-FilterValues = Union[int, float, str, bool, datetime.date, None, FilterValuesList]
-
-
-@dataclass
-class _FilterValue(_Filters):
-    path: Union[str, List[str]]
-    value: FilterValues
-    operator: weaviate_pb2.Filters.Operator
-
-    def __and__(self, other: "_Filters"):
-        return _FilterAnd(self, other)
-
-    def __or__(self, other: "_Filters"):
-        return _FilterOr(self, other)
-
-
-@dataclass
-class Filter:
-    path: Union[str, List[str]]
-    length: bool = False
-
-    def __post_init__(self):
-        if isinstance(self.path, str):
-            path = [self.path]
-        else:
-            path = self.path
-
-        if self.length:
-            self.__internal_path = "len(" + path[-1] + ")"
-        else:
-            self.__internal_path = path
-
-    def is_none(self, val: bool) -> _FilterValue:
-        return _FilterValue(
-            path=self.__internal_path, value=val, operator=weaviate_pb2.Filters.OPERATOR_IS_NULL
-        )
-
-    def contains_any(self, val: FilterValuesList) -> _FilterValue:
-        return _FilterValue(
-            path=self.__internal_path,
-            value=val,
-            operator=weaviate_pb2.Filters.OPERATOR_CONTAINS_ANY,
-        )
-
-    def contains_all(self, val: FilterValuesList) -> _FilterValue:
-        return _FilterValue(
-            path=self.__internal_path,
-            value=val,
-            operator=weaviate_pb2.Filters.OPERATOR_CONTAINS_ALL,
-        )
-
-    def equal(self, val: FilterValues) -> _FilterValue:
-        return _FilterValue(
-            path=self.__internal_path, value=val, operator=weaviate_pb2.Filters.OPERATOR_EQUAL
-        )
-
-    def not_equal(self, val: FilterValues) -> _FilterValue:
-        return _FilterValue(
-            path=self.__internal_path, value=val, operator=weaviate_pb2.Filters.OPERATOR_NOT_EQUAL
-        )
-
-    def less_than(self, val: FilterValues) -> _FilterValue:
-        return _FilterValue(
-            path=self.__internal_path, value=val, operator=weaviate_pb2.Filters.OPERATOR_LESS_THAN
-        )
-
-    def less_than_equal(self, val: FilterValues) -> _FilterValue:
-        return _FilterValue(
-            path=self.__internal_path,
-            value=val,
-            operator=weaviate_pb2.Filters.OPERATOR_LESS_THAN_EQUAL,
-        )
-
-    def greater_than(self, val: FilterValues) -> _FilterValue:
-        return _FilterValue(
-            path=self.__internal_path,
-            value=val,
-            operator=weaviate_pb2.Filters.OPERATOR_GREATER_THAN,
-        )
-
-    def greater_than_equal(self, val: FilterValues) -> _FilterValue:
-        return _FilterValue(
-            path=self.__internal_path,
-            value=val,
-            operator=weaviate_pb2.Filters.OPERATOR_GREATER_THAN_EQUAL,
-        )
-
-    def like(self, val: str) -> _FilterValue:
-        return _FilterValue(
-            path=self.__internal_path, value=val, operator=weaviate_pb2.Filters.OPERATOR_LIKE
-        )
