+++ weaviate/collection/collection_base.py
@@ -2,7 +2,7 @@ from typing import Dict
 
 from requests.exceptions import ConnectionError as RequestsConnectionError
 
-from weaviate.collection.classes import (
+from weaviate.collection.classes.config import (
     CollectionConfigCreateBase,
     _collection_configs_from_json,
     _CollectionConfig,
